+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.0] - 2021-12-20
+
+- show timer in status bar
+- click on status bar to cancel or reset timer
+- change commands
+
 ## [0.1.2] - 2020-07-13
 
 ### Add